ARM: dts: exynos4412-odroid-common: separate buck8 regulator
Move the buck8 regulator to the respective board dts files, since
it is used differently depending on the board.

On the ODROID-X2 buck8 is used as supply for the eMMC, with the
vendor kernel setting it up with a voltage of 2.85V. On the U2/U3
it is the supply for the LAN9730 block on the board, which
provides the ethernet functionality. The vendor kernel sets
a voltage for 3.3V in the case.